+++ doc/test/run.t
@@ -6,9 +6,7 @@ When changing Dune version, you need to update the docs too to make this test pa
 Occasionally we do want to mention an older Dune version in documentation. This
 is fine, but you then need to update the list of such exceptions below.
 
-  $ touch dune
-  $ dune build > /dev/null 2> /dev/null
-  $ DUNE_LANG=$(cat dune-project)
+  $ DUNE_LANG=$(dune compute latest-lang-version | sed 's/"//g')
   $ grep '(lang dune' ../*.rst | grep -v "$DUNE_LANG"
   ../formatting.rst:If using ``(lang dune 2.0)``, there is nothing to setup in dune, formatting will
   ../formatting.rst:.. note:: This section applies only to projects with ``(lang dune 1.x)``.