+++ test/blackbox-tests/test-cases/mdx-stanza.t/run.t
@@ -7,7 +7,12 @@ dune-project
 
   $ dune build @install --root using-mdx/
   Entering directory 'using-mdx'
-  Info: Appending this line to dune-project: (using mdx 0.1)
+  File "dune", line 1, characters 0-5:
+  1 | (mdx)
+      ^^^^^
+  Error: 'mdx' is available only when mdx is enabled in the dune-project file.
+  You must enable it using (using mdx 0.1) in your dune-project file.
+  [1]
 
 It also requires dune lang 2.4 or higher